Add checks if files exist

+++ synapse/data/start.sh
@@ -1,4 +1,5 @@
-#!/bin/bash
-rm -rf /data/media_store
-rm -f /data/homeserver.db
+#!/bin/sh
+
+[ -d /data/media_store ] && rm -rf /data/media_store
+[ -f /data/homeserver.db ] && rm -f /data/homeserver.db
 python /start.py
